requirements.txt
dask==2022.12.0
gunpowder==1.2.2
ipympl==0.8.8
ipython==8.7.0
ipywidgets==7.7.2
jupyter==1.0.0
jupyter_core==4.8.1
Keras==2.3.1
matplotlib==3.3.4
natsort==8.1.0
nose==1.3.7
numcodecs
numpy==1.21.0
opencv-python==4.2.0.32
pandas==1.1.5
protobuf==3.20.1
pydot==1.4.1
PyYAML>=5.4
scikit-image==0.17.2
scikit-learn==0.21.2
scipy==1.2.1
tensorflow==1.13.1
testfixtures==6.7.0
tqdm
zarr
